+++ CHANGELOG.md
@@ -4,9 +4,9 @@ All notable changes to this project will be documented in this file.
 
 ## [2.3] - Unreleased
 
-- This release introduces the fixes necessary for compatibility with
-  [libcommute 1.0.0](https://github.com/krivenko/libcommute/releases/tag/v1.0.0)
-  .
+- Added new method `Hamiltonian::getNumBlocks()`.
+
+- Bumped required libcommute version to 1.0.0.
 
 ## [2.2] - 2025-09-08